rename: RedisUtil을 RedisStringUtil로 변경하고 common/ 으로 이동
